LPS-179975 Refactoring Policy Activities Details
