--- modules/apps/site-initializer/site-initializer-raylife-ap/extra/remote-app/src/routes/policies/pages/components/Activities/index.scss
@@ -1,147 +1,0 @@
-.table-list-container {
-	.border-link-active::after {
-		background-color: var(--primary);
-		content: '';
-		display: block;
-		height: 1px;
-		margin-top: 6px;
-		position: absolute;
-		width: 25px;
-	}
-
-	.blue-line-activites:after {
-		background-color: #4c84ff;
-		bottom: -1px;
-		content: '';
-		display: block;
-		left: -1px;
-		position: absolute;
-		top: 0;
-		width: 5px;
-	}
-
-	.box-activites {
-		height: 53px;
-	}
-
-	.box-table {
-		tbody {
-			display: flex;
-			flex-direction: column;
-			height: 230px;
-			overflow: auto;
-			width: 100%;
-		}
-
-		thead {
-			display: flex;
-		}
-
-		thead tr {
-			display: flex;
-			justify-content: space-between;
-			padding-right: 90px;
-			width: 100%;
-		}
-		tbody tr {
-			display: flex;
-			justify-content: space-between;
-			width: 100%;
-		}
-	}
-
-	.box-table-mobile {
-		thead {
-			display: table;
-			width: 100%;
-		}
-		tbody tr {
-			justify-content: space-between;
-			width: 100%;
-		}
-	}
-
-	.box-shadow {
-		box-shadow: 0 5px 20px rgb(0 0 0 / 10%);
-		z-index: 1;
-	}
-
-	.border-line-selected:after {
-		background-color: #4c84ff;
-		bottom: -1px;
-		content: '';
-		display: block;
-		left: -1px;
-		position: absolute;
-		top: 0;
-		width: 5px;
-	}
-
-	.dotted-line :before {
-		background-image: linear-gradient(
-			to right,
-			#aaa 0%,
-			#aaa 20%,
-			transparent 20%
-		);
-		background-position: 0 0;
-		background-repeat: repeat-x;
-		background-size: 6px 1px;
-		bottom: 0;
-		content: '';
-		height: 1px;
-		left: 10px;
-		position: absolute;
-		right: 10px;
-	}
-	.border {
-		color: #272833;
-		font-size: 0.75rem;
-		font-weight: 700;
-
-		&.border-header {
-			border-top: 1px solid #ccdae8;
-		}
-		&.border-bottom {
-			border-bottom: 1px solid #ccdae8;
-		}
-	}
-
-	.font {
-		font-family: 'Barlow', system-ui, -apple-system, blinkmacsystemfont,
-			'Segoe UI', roboto, oxygen-sans, ubuntu, cantarell, 'Helvetica Neue',
-			arial, sans-serif, 'Apple Color Emoji', 'Segoe UI Emoji',
-			'Segoe UI Symbol';
-
-		font-size: 1em;
-		font-weight: 600;
-		line-height: 1.5;
-	}
-
-	.font-table {
-		color: #6b6c7e;
-		font-size: 0.815rem;
-		font-weight: 500;
-	}
-
-	.font-table-bold {
-		font-size: 0.815rem;
-		font-weight: bold;
-	}
-
-	@media screen and (max-width: 1030px) {
-		.font-table,
-		.font-table-bold {
-			font-size: 0.835rem;
-			margin-right: 6rem;
-		}
-	}
-
-	@media screen and (max-width: 500px) {
-		.font-table,
-		.font-table-bold {
-			font-size: 0.775rem;
-			margin-right: 2rem;
-		}
-	}
-}
